Break out ExamSegmentWrapper and ExamPageWrapper resources into their own controllers, services, and repositories.
