+++ service/src/main/java/tds/exam/services/impl/ExamSegmentServiceImpl.java
@@ -7,10 +7,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
@@ -33,8 +31,6 @@ import tds.exam.services.ExamSegmentService;
 import tds.exam.services.FieldTestService;
 import tds.exam.services.FormSelector;
 import tds.exam.services.SegmentPoolService;
-import tds.exam.wrapper.ExamPageWrapper;
-import tds.exam.wrapper.ExamSegmentWrapper;
 
 @Service
 public class ExamSegmentServiceImpl implements ExamSegmentService {
@@ -204,36 +200,5 @@ public class ExamSegmentServiceImpl implements ExamSegmentService {
     @Override
     public boolean checkIfSegmentsCompleted(final UUID examId) {
         return examSegmentQueryRepository.findCountOfUnsatisfiedSegments(examId) == 0;
-    }
-
-    @Override
-    public List<ExamSegmentWrapper> findAllExamSegments(final UUID examId) {
-        Map<String, List<ExamPageWrapper>> examPageWrappersBySegmentKey = examPageService.findPagesWithItems(examId)
-            .stream()
-            .collect(Collectors.groupingBy(examPageWrapper -> examPageWrapper.getExamPage().getSegmentKey()));
-
-        return examSegmentQueryRepository.findByExamId(examId)
-            .stream()
-            .map(examSegment -> new ExamSegmentWrapper(examSegment, examPageWrappersBySegmentKey.get(examSegment.getSegmentKey())))
-            .collect(Collectors.toList());
-    }
-
-    @Override
-    public Optional<ExamSegmentWrapper> findExamSegment(final UUID examId, final int segmentPosition) {
-        Optional<ExamSegment> maybeExamSegment = examSegmentQueryRepository.findByExamIdAndSegmentPosition(examId, segmentPosition);
-
-        return maybeExamSegment.map(examSegment -> new ExamSegmentWrapper(examSegment, examPageService.findPagesForExamSegment(examId, examSegment.getSegmentKey())));
-    }
-
-    @Override
-    public Optional<ExamSegmentWrapper> findExamSegmentWithPageAtPosition(final UUID examId, final int segmentPosition, final int pagePosition) {
-        Optional<ExamSegment> maybeExamSegment = examSegmentQueryRepository.findByExamIdAndSegmentPosition(examId, segmentPosition);
-        Optional<ExamPageWrapper> maybeExamPageWrapper = examPageService.findPageWithItems(examId, pagePosition);
-
-        if (!maybeExamSegment.isPresent() || !maybeExamPageWrapper.isPresent()) {
-            return Optional.empty();
-        }
-
-        return Optional.of(new ExamSegmentWrapper(maybeExamSegment.get(), Collections.singletonList(maybeExamPageWrapper.get())));
     }
 }
