+++ service/src/main/java/tds/exam/web/endpoints/ExamSegmentController.java
@@ -8,7 +8,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
@@ -16,12 +15,10 @@ import java.util.Optional;
 import java.util.UUID;
 
 import tds.common.ValidationError;
-import tds.common.web.exceptions.NotFoundException;
 import tds.common.web.resources.NoContentResponseResource;
 import tds.exam.ExamSegment;
 import tds.exam.services.ExamSegmentService;
 import tds.exam.web.annotations.VerifyAccess;
-import tds.exam.wrapper.ExamSegmentWrapper;
 
 @RestController
 @RequestMapping("/exam")
@@ -55,33 +52,5 @@ public class ExamSegmentController {
         }
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
-
-    @GetMapping(value = "{examId}/segmentWrappers", produces = MediaType.APPLICATION_JSON_VALUE)
-    ResponseEntity<List<ExamSegmentWrapper>> findExamSegmentWrappersForExam(@PathVariable final UUID examId) {
-        List<ExamSegmentWrapper> wrappers = examSegmentService.findAllExamSegments(examId);
-
-        if (wrappers.isEmpty()) {
-            throw new NotFoundException("Could not find exam segments for exam %s", examId);
-        }
-
-        return ResponseEntity.ok(wrappers);
-    }
-
-    @GetMapping(value = "{examId}/segmentWrappers/{segmentPosition}", produces = MediaType.APPLICATION_JSON_VALUE)
-    ResponseEntity<ExamSegmentWrapper> findExamSegmentWrapperForExamAndSegmentPosition(@PathVariable final UUID examId,
-                                                                                       @PathVariable final int segmentPosition,
-                                                                                       @RequestParam(value = "pagePosition", required = false) Integer pagePosition) {
-        if (pagePosition == null) {
-            ExamSegmentWrapper examSegmentWrapper = examSegmentService.findExamSegment(examId, segmentPosition)
-                .orElseThrow(() -> new NotFoundException("Could not find exam segment for exam %s and position %d", examId, segmentPosition));
-
-            return ResponseEntity.ok(examSegmentWrapper);
-        }
-
-        ExamSegmentWrapper examSegmentWrapper = examSegmentService.findExamSegmentWithPageAtPosition(examId, segmentPosition, pagePosition)
-            .orElseThrow(() -> new NotFoundException("Could not find exam segment for exam %s, segment position %d, and page position %d", examId, segmentPosition, pagePosition));
-
-        return ResponseEntity.ok(examSegmentWrapper);
     }
 }
